requirements.txt
requests==2.28.2
Flask==2.2.2
Flask-RESTful==0.3.10
gunicorn==20.1.0
Flask-Cors==3.0.10
Werkzeug==2.2.2
python-dotenv==1.0.0